Working MSC code for clock image.
In this MSC implementation I am able to locate the location of alarm
clock in the real image. This implementation just takes into account one
rotation and is able to correctly locate alarm clock in the real image.
To implement edge detection I am performing image segmentation first in
the memory image as well as in the test image.