No changes made in the code. Working for scaling.
No changes made in the code. Working for scaling.

Use the getPointsForRotation function in the original code. Because we
have to be very careful for almost each and every points on which we
will be using backslash. Otherwise everything might screw up. Now try
with moving the axes or shifting the origin.