Layer differentiation remains to be a stickler.
It seems the layer differentiation still remains to be a stickler. I
guess I might have to move back to row by row rank calculation to see
whether I am able to differentiate among independent functions or layers
based on that.
Earlier I was trying to get rank based on column-by-column basis, but
the correct method would be to check rank on row-by-row basis.