Able to learn new images and store them.
Able to learn new images and store them. Currently I’ve tried with a
bell pepper image that was stored and tried learning a rotated sailboat
image. When I saw that bell pepper does not image perfectly with
sailboat image I learnt the given sailboat image and stored it in the
memory.